# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Aptum Scheduling",
    "description" : "Terminsuche und Buchung für Physio- und Ergotherapiepraxen. Jeder Vorschlag und jede Buchung läuft durch dasselbe deterministische Regelwerk der Heilmittel-Richtlinie; die Antworten nennen je Regel Ausgang und Begründung. Daten sind je Mandant getrennt (Row Level Security).",
    "version" : "0.1"
  },
  "paths" : {
    "/verordnungen" : {
      "post" : {
        "tags" : [ "verordnung-controller" ],
        "operationId" : "anlegen",
        "parameters" : [ {
          "name" : "X-Mandant",
          "in" : "header",
          "description" : "Die Praxis, für die die Anfrage gilt. Platzhalter für Authentifizierung: In einer echten Anwendung käme der Mandant aus einem signierten Token.",
          "required" : true,
          "example" : "praxis-a"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VerordnungAnlage"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fehler"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fehler"
                }
              }
            }
          },
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VerordnungAngelegt"
                }
              }
            }
          }
        }
      }
    },
...
    "/verordnungen/{id}" : {
      "get" : {
        "tags" : [ "verordnung-controller" ],
        "operationId" : "ansehen",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "X-Mandant",
          "in" : "header",
          "description" : "Die Praxis, für die die Anfrage gilt. Platzhalter für Authentifizierung: In einer echten Anwendung käme der Mandant aus einem signierten Token.",
          "required" : true,
          "example" : "praxis-a"
        } ],
        "responses" : {
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fehler"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Fehler"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Verordnungsakte"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Fehler" : {
        "type" : "object",
        "properties" : {
          "fehler" : {
            "type" : "string"
          }
        }
      },
      "VerordnungAnlage" : {
        "type" : "object",
        "properties" : {
          "ausstellungsdatum" : {
            "type" : "string",
            "format" : "date"
          },
          "dringlicherBedarf" : {
            "type" : "boolean"
          },
          "diagnosegruppe" : {
            "type" : "string"
          },
          "verordneteEinheiten" : {
            "type" : "integer",
            "format" : "int32"
          },
          "frequenzMin" : {
            "type" : "integer",
            "format" : "int32"
          },
          "frequenzMax" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "VerordnungAngelegt" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
...
      "Regel" : {
        "type" : "object",
        "properties" : {
          "regel" : {
            "type" : "string"
          },
          "ausgang" : {
            "type" : "string"
          },
          "begruendung" : {
            "type" : "string"
          }
        }
      },
...
      "Abrechnungsposten" : {
        "type" : "object",
        "properties" : {
          "verordnung" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ausstellungsdatum" : {
            "type" : "string",
            "format" : "date"
          },
          "diagnosegruppe" : {
            "type" : "string"
          },
          "therapieform" : {
            "type" : "string"
          },
          "verordnet" : {
            "type" : "integer",
            "format" : "int32"
          },
          "erbracht" : {
            "type" : "integer",
            "format" : "int32"
          },
          "offen" : {
            "type" : "integer",
            "format" : "int32"
          },
          "ersteBehandlung" : {
            "type" : "string",
            "format" : "date"
          },
          "letzteBehandlung" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string"
          },
          "begruendung" : {
            "type" : "string"
          },
          "regeln" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Regel"
            }
          }
        }
      },
      "Verordnungsakte" : {
        "type" : "object",
        "properties" : {
          "posten" : {
            "$ref" : "#/components/schemas/Abrechnungsposten"
          },
          "dringlicherBedarf" : {
            "type" : "boolean"
          },
          "frequenzMin" : {
            "type" : "integer",
            "format" : "int32"
          },
          "frequenzMax" : {
            "type" : "integer",
            "format" : "int32"
          },
          "frequenz" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}